[Feature] Added data prepping and visualisation, elbow method to select k clusters.
Added data preparing step which analyses the data and uses a min max scaler to normalize the data.

Created visualisations for viewing the existing clusters (since we have the labels)

Added the elbow method implementation to automatically select the optimal k value, and overrode it to the next value because we know the classes.